Random test fixes - fixed typo in constants.ts - fixed mocha type recognition - refactored test to use chai-as-promised
